refactor: move calls to LLM into CdGpt services
Moved the functions getAnalysisContent and getChatContent from the models
into a service layer above the model. The reasoning for this is to
have a call to an external API encapsulated in a module that's separate
from the EntryAnalysis and EntryConversation models so that they
don't depend on it.

Unit tests for these functions were also moved into a new test file,
models/services/CdGpt.test.ts
